refactor: split test files into separate test suites
Split the merged test-prompt-sub-agents.mjs into two separate files:
- test-prompt-plan-sub-agent.mjs (5 tests for --prompt-plan-sub-agent)
- test-prompt-general-purpose-sub-agents.mjs (6 tests for --prompt-general-purpose-sub-agent)

Both test files run independently and all tests pass.
